Moved the workflow into its own area so that it is easier when importing the code for other tools
