+++ process_test.py
@@ -20,63 +20,8 @@ from __future__ import print_function
 
 import argparse
 
-from basic_modules.workflow import Workflow
 from utils import logger
-
-from mg_process_test.tool.testTool import testTool
-
-# ------------------------------------------------------------------------------
-
-
-class process_test(Workflow):
-    """
-    Functions for demonstrating the pipeline set up.
-    """
-
-    configuration = {}
-
-    def __init__(self, configuration=None):
-        """
-        Initialise the tool with its configuration.
-
-        Parameters
-        ----------
-        configuration : dict
-           a dictionary containing parameters that define how the operation
-           should be carried out, which are specific to each Tool.
-        """
-        logger.info("Processing Test")
-        if configuration is None:
-            configuration = {}
-
-        self.configuration.update(configuration)
-
-    def run(self, input_files, metadata, output_files):
-        """
-        Main run function for processing a test file.
-
-        Parameters
-        ----------
-        input_files : dict
-           Dictionary of file locations
-        metadata : list
-           Required meta data
-        output_files : dict
-           Locations of the output files to be returned by the pipeline
-
-        Returns
-        -------
-        output_files : dict
-           Locations for the output txt
-        output_metadata : dict
-           Matching metadata for each of the files
-        """
-
-        # Initialise the test tool
-        tt_handle = testTool(self.configuration)
-        tt_files, tt_meta = tt_handle.run(input_files, metadata, output_files)
-
-        return (tt_files, tt_meta)
+from mg_process_test.workflow.test_workflow import process_test
 
 # ------------------------------------------------------------------------------
 
@@ -104,6 +49,7 @@ def main_json(config, in_metadata, out_metadata):
     return result
 
 # ------------------------------------------------------------------------------
+
 
 if __name__ == "__main__":
